Migrate Device System API (/system/device) to Service Information API.
