+++ dConnectDevicePlugin/dConnectDeviceIrKit/src/org/deviceconnect/android/deviceplugin/irkit/profile/IRKitSystemProfile.java
@@ -6,12 +6,9 @@
  */
 package org.deviceconnect.android.deviceplugin.irkit.profile;
 
-import org.deviceconnect.android.deviceplugin.irkit.IRKitDevice;
-import org.deviceconnect.android.deviceplugin.irkit.IRKitDeviceService;
 import org.deviceconnect.android.deviceplugin.irkit.settings.activity.IRKitSettingActivity;
 import org.deviceconnect.android.event.EventManager;
 import org.deviceconnect.android.message.MessageUtils;
-import org.deviceconnect.android.profile.DConnectProfileProvider;
 import org.deviceconnect.android.profile.SystemProfile;
 import org.deviceconnect.message.DConnectMessage;
 
@@ -25,33 +22,11 @@ import android.os.Bundle;
  */
 public class IRKitSystemProfile extends SystemProfile {
 
-    /**
-     * System Profile.
-     * 
-     * @param provider Provider
-     */
-    public IRKitSystemProfile(final DConnectProfileProvider provider) {
-        super(provider);
-    }
-
     @Override
     protected Class<? extends Activity> getSettingPageActivity(final Intent request, final Bundle param) {
         return IRKitSettingActivity.class;
     }
 
-    @Override
-    protected ConnectState getWifiState(final String serviceId) {
-        
-        IRKitDeviceService service = (IRKitDeviceService) getContext();
-        IRKitDevice device = service.getDevice(serviceId);
-        
-        if (device != null) {
-            return ConnectState.ON;
-        } else {
-            return ConnectState.OFF;
-        }
-    }
-    
     @Override
     protected boolean onDeleteEvents(final Intent request, final Intent response, final String sessionKey) {
