+++ dConnectDevicePlugin/dConnectDevicePluginSDK/src/org/deviceconnect/android/profile/ServiceInformationProfile.java
@@ -1,0 +1,338 @@
+/*
+ SystemProfile.java
+ Copyright (c) 2015 NTT DOCOMO,INC.
+ Released under the MIT license
+ http://opensource.org/licenses/mit-license.php
+ */
+package org.deviceconnect.android.profile;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import org.deviceconnect.android.message.MessageUtils;
+import org.deviceconnect.message.DConnectMessage;
+import org.deviceconnect.profile.ServiceInformationProfileConstants;
+
+import android.content.Intent;
+import android.content.pm.PackageInfo;
+import android.content.pm.PackageManager;
+import android.content.pm.PackageManager.NameNotFoundException;
+import android.os.Bundle;
+
+/**
+ * Service Information プロファイル.
+ * 
+ * <p>
+ * サービス情報を提供するAPI.<br/>
+ * サービス情報を提供するデバイスプラグインは当クラスを継承し、対応APIを実装すること。 <br/>
+ * </p>
+ * 
+ * <h1>各API提供メソッド</h1>
+ * <p>
+ * System Profile の各APIへのリクエストに対し、以下のコールバックメソッド群が自動的に呼び出される。<br/>
+ * サブクラスは以下のメソッド群からデバイスプラグインが提供するAPI用のメソッドをオーバーライドし、機能を実装すること。<br/>
+ * オーバーライドされていない機能は自動的に非対応APIとしてレスポンスを返す。
+ * </p>
+ * <ul>
+ * <li>System API [GET] :
+ * {@link ServiceInformationProfile#onGetInformation(Intent, Intent, String)}</li>
+ * </ul>
+ * 
+ * @author NTT DOCOMO, INC.
+ */
+public abstract class ServiceInformationProfile extends DConnectProfile implements ServiceInformationProfileConstants {
+
+    /**
+     * 設定画面起動用IntentのパラメータオブジェクトのExtraキー.
+     */
+    public static final String SETTING_PAGE_PARAMS = "org.deviceconnect.profile.system.setting_params";
+
+    /**
+     * プロファイルプロバイダー.
+     */
+    private DConnectProfileProvider mProvider;
+
+    /**
+     * 指定されたプロファイルプロバイダーをもつSystemプロファイルを生成する.
+     * 
+     * @param provider プロファイルプロバイダー
+     */
+    public ServiceInformationProfile(final DConnectProfileProvider provider) {
+        this.mProvider = provider;
+    }
+
+    /**
+     * プロファイルプロバイダーを取得する.
+     * 
+     * @return プロファイルプロバイダー
+     */
+    protected DConnectProfileProvider getProfileProvider() {
+        return mProvider;
+    }
+
+    @Override
+    public final String getProfileName() {
+        return PROFILE_NAME;
+    }
+
+    /**
+     * AndroidManifest.xmlのversionNameを取得する.
+     * 
+     * @return バージョン名
+     */
+    private String getCurrentVersionName() {
+        PackageManager packageManager = getContext().getPackageManager();
+        try {
+            PackageInfo packageInfo = packageManager.getPackageInfo(getContext().getPackageName(),
+                    PackageManager.GET_ACTIVITIES);
+            return packageInfo.versionName;
+        } catch (NameNotFoundException e) {
+            return "Unknown";
+        }
+    }
+
+    /**
+     * WiFiの接続状態を取得する.
+     * 
+     * @param serviceId サービスID
+     * @return WiFiの接続状態
+     * @see ConnectState
+     */
+    protected ConnectState getWifiState(final String serviceId) {
+        return ConnectState.NONE;
+    }
+
+    /**
+     * Bluetoothの接続状態を取得する.
+     * 
+     * @param serviceId サービスID
+     * @return Bluetoothの接続状態
+     * @see ConnectState
+     */
+    protected ConnectState getBluetoothState(final String serviceId) {
+        return ConnectState.NONE;
+    }
+
+    /**
+     * NFCの接続状態を取得する.
+     * 
+     * @param serviceId サービスID
+     * @return NFCの接続状態
+     * @see ConnectState
+     */
+    protected ConnectState getNFCState(final String serviceId) {
+        return ConnectState.NONE;
+    }
+
+    /**
+     * BLEの接続状態を取得する.
+     * 
+     * @param serviceId サービスID
+     * @return BLEの接続状態
+     * @see ConnectState
+     */
+    protected ConnectState getBLEState(final String serviceId) {
+        return ConnectState.NONE;
+    }
+
+    @Override
+    protected boolean onGetRequest(final Intent request, final Intent response) {
+        String attribute = getAttribute(request);
+        boolean result = true;
+        String serviceId = getServiceID(request);
+
+        if (attribute == null) {
+            result = onGetInformation(request, response, serviceId);
+        } else {
+            MessageUtils.setUnknownAttributeError(response);
+        }
+
+        return result;
+    }
+
+    // ------------------------------------
+    // GET
+    // ------------------------------------
+
+    /**
+     * 周辺機器のサービス情報取得リクエストハンドラー.<br/>
+     * 周辺機器のサービス情報取得を提供し、その結果をレスポンスパラメータに格納する。
+     * レスポンスパラメータの送信準備が出来た場合は返り値にtrueを指定する事。
+     * 送信準備ができていない場合は、返り値にfalseを指定し、スレッドを立ち上げてそのスレッドで最終的にレスポンスパラメータの送信を行う事。
+     * このメソッドでは自動的にサービス情報を返信する。返信処理に変更を加えたい場合はオーバーライドし、処理を上書きすること。
+     * 
+     * @param request リクエストパラメータ
+     * @param response レスポンスパラメータ
+     * @param serviceId サービスID
+     * @return レスポンスパラメータを送信するか否か
+     */
+    protected boolean onGetInformation(final Intent request, final Intent response, final String serviceId) {
+
+        // connect
+        Bundle connect = new Bundle();
+        setWifiState(connect, getWifiState(serviceId));
+        setBluetoothState(connect, getBluetoothState(serviceId));
+        setNFCState(connect, getNFCState(serviceId));
+        setBLEState(connect, getBLEState(serviceId));
+        setConnect(response, connect);
+
+        // version
+        setVersion(response, getCurrentVersionName());
+
+        // supports
+        ArrayList<String> profiles = new ArrayList<String>();
+        for (DConnectProfile profile : mProvider.getProfileList()) {
+            profiles.add(profile.getProfileName());
+        }
+        setSupports(response, profiles.toArray(new String[0]));
+        setResult(response, DConnectMessage.RESULT_OK);
+
+        return true;
+    }
+
+    // ------------------------------------
+    // レスポンスセッターメソッド群
+    // ------------------------------------
+
+    /**
+     * レスポンスにバージョンを格納する.
+     * 
+     * @param response レスポンスパラメータ
+     * @param version バージョン
+     */
+    public static void setVersion(final Intent response, final String version) {
+        response.putExtra(PARAM_VERSION, version);
+    }
+
+    /**
+     * レスポンスにサポートしているI/Fの一覧を格納する.
+     * 
+     * @param response レスポンスパラメータ
+     * @param supports サポートしているI/F一覧
+     */
+    public static void setSupports(final Intent response, final String[] supports) {
+        response.putExtra(PARAM_SUPPORTS, supports);
+    }
+
+    /**
+     * レスポンスにサポートしているI/Fの一覧を格納する.
+     * 
+     * @param response レスポンスパラメータ
+     * @param supports サポートしているI/F一覧
+     */
+    public static void setSupports(final Intent response, final List<String> supports) {
+        setSupports(response, supports.toArray(new String[supports.size()]));
+    }
+
+    /**
+     * レスポンスにデバイスの接続状態を設定する.
+     * 
+     * @param response レスポンスパラメータ
+     * @param connect デバイスの接続状態
+     */
+    public static void setConnect(final Intent response, final Bundle connect) {
+        response.putExtra(PARAM_CONNECT, connect);
+    }
+
+    /**
+     * 指定されたパラメータに接続状態を設定する.
+     * 
+     * @param b バンドルオブジェクト
+     * @param key パラメータキー
+     * @param state 状態
+     */
+    private static void setConnectionState(final Bundle b, final String key, final ConnectState state) {
+        // 非対応のものは省略
+        switch (state) {
+        case ON:
+            b.putBoolean(key, true);
+            break;
+        case OFF:
+            b.putBoolean(key, false);
+            break;
+        default:
+            break;
+        }
+    }
+
+    /**
+     * デバイスプラグインの接続状態にWiFiの接続状態を設定する.
+     * 
+     * @param connect デバイスプラグインの接続状態パラメータ
+     * @param connecting true:ON、false:OFF
+     */
+    public static void setWifiState(final Bundle connect, final boolean connecting) {
+        connect.putBoolean(PARAM_WIFI, connecting);
+    }
+
+    /**
+     * デバイスプラグインの接続状態にWiFiの接続状態を設定する.
+     * 
+     * @param connect デバイスプラグインの接続状態パラメータ
+     * @param state 接続状態
+     */
+    public static void setWifiState(final Bundle connect, final ConnectState state) {
+        setConnectionState(connect, PARAM_WIFI, state);
+    }
+
+    /**
+     * デバイスプラグインの接続状態にBluetoothの接続状態を設定する.
+     * 
+     * @param connect デバイスプラグインの接続状態パラメータ
+     * @param connecting true:ON、false:OFF
+     */
+    public static void setBluetoothState(final Bundle connect, final boolean connecting) {
+        connect.putBoolean(PARAM_BLUETOOTH, connecting);
+    }
+
+    /**
+     * デバイスプラグインの接続状態にBluetoothの接続状態を設定する.
+     * 
+     * @param connect デバイスプラグインの接続状態パラメータ
+     * @param state 接続状態
+     */
+    public static void setBluetoothState(final Bundle connect, final ConnectState state) {
+        setConnectionState(connect, PARAM_BLUETOOTH, state);
+    }
+
+    /**
+     * デバイスプラグインの接続状態にNFCの接続状態を設定する.
+     * 
+     * @param connect デバイスプラグインの接続状態パラメータ
+     * @param connecting true:ON、false:OFF
+     */
+    public static void setNFCState(final Bundle connect, final boolean connecting) {
+        connect.putBoolean(PARAM_NFC, connecting);
+    }
+
+    /**
+     * デバイスプラグインの接続状態にNFCの接続状態を設定する.
+     * 
+     * @param connect デバイスプラグインの接続状態パラメータ
+     * @param state 接続状態
+     */
+    public static void setNFCState(final Bundle connect, final ConnectState state) {
+        setConnectionState(connect, PARAM_NFC, state);
+    }
+
+    /**
+     * デバイスプラグインの接続状態にBLEの接続状態を設定する.
+     * 
+     * @param connect デバイスプラグインの接続状態パラメータ
+     * @param connecting true:ON、false:OFF
+     */
+    public static void setBLEState(final Bundle connect, final boolean connecting) {
+        connect.putBoolean(PARAM_BLE, connecting);
+    }
+
+    /**
+     * デバイスプラグインの接続状態にBLEの接続状態を設定する.
+     * 
+     * @param connect デバイスプラグインの接続状態パラメータ
+     * @param state 接続状態
+     */
+    public static void setBLEState(final Bundle connect, final ConnectState state) {
+        setConnectionState(connect, PARAM_BLE, state);
+    }
+
+}
