+++ dConnectDevicePlugin/dConnectDevicePluginSDK/src/org/deviceconnect/android/profile/SystemProfile.java
@@ -6,19 +6,14 @@
  */
 package org.deviceconnect.android.profile;
 
-import java.util.ArrayList;
 import java.util.List;
 
 import org.deviceconnect.android.message.MessageUtils;
 import org.deviceconnect.message.DConnectMessage;
 import org.deviceconnect.profile.SystemProfileConstants;
-import org.deviceconnect.profile.SystemProfileConstants.ConnectState;
 
 import android.app.Activity;
 import android.content.Intent;
-import android.content.pm.PackageInfo;
-import android.content.pm.PackageManager;
-import android.content.pm.PackageManager.NameNotFoundException;
 import android.os.Bundle;
 
 /**
@@ -38,8 +33,6 @@ import android.os.Bundle;
  * <ul>
  * <li>System API [GET] :
  * {@link SystemProfile#onGetSystem(Intent, Intent, String)}</li>
- * <li>Device System API[GET] :
- * {@link SystemProfile#onGetDevice(Intent, Intent, String)}</li>
  * <li>Device System Wake Up API [PUT] :
  * {@link SystemProfile#onPutWakeup(Intent, Intent, String)}</li>
  * <li>Device System Wake Up API [DELETE] :
@@ -56,20 +49,6 @@ public abstract class SystemProfile extends DConnectProfile implements SystemPro
     public static final String SETTING_PAGE_PARAMS = "org.deviceconnect.profile.system.setting_params";
 
     /**
-     * プロファイルプロバイダー.
-     */
-    private DConnectProfileProvider mProvider;
-
-    /**
-     * 指定されたプロファイルプロバイダーをもつSystemプロファイルを生成する.
-     * 
-     * @param provider プロファイルプロバイダー
-     */
-    public SystemProfile(final DConnectProfileProvider provider) {
-        this.mProvider = provider;
-    }
-
-    /**
      * 遷移先の設定画面用のActivityのクラス.
      * 
      * @param request リクエストパラメータ
@@ -81,78 +60,9 @@ public abstract class SystemProfile extends DConnectProfile implements SystemPro
      */
     protected abstract Class<? extends Activity> getSettingPageActivity(Intent request, Bundle param);
 
-    /**
-     * プロファイルプロバイダーを取得する.
-     * 
-     * @return プロファイルプロバイダー
-     */
-    protected DConnectProfileProvider getProfileProvider() {
-        return mProvider;
-    }
-
     @Override
     public final String getProfileName() {
         return PROFILE_NAME;
-    }
-
-    /**
-     * AndroidManifest.xmlのversionNameを取得する.
-     * 
-     * @return バージョン名
-     */
-    private String getCurrentVersionName() {
-        PackageManager packageManager = getContext().getPackageManager();
-        try {
-            PackageInfo packageInfo = packageManager.getPackageInfo(getContext().getPackageName(),
-                    PackageManager.GET_ACTIVITIES);
-            return packageInfo.versionName;
-        } catch (NameNotFoundException e) {
-            return "Unknown";
-        }
-    }
-
-    /**
-     * WiFiの接続状態を取得する.
-     * 
-     * @param serviceId サービスID
-     * @return WiFiの接続状態
-     * @see ConnectState
-     */
-    protected ConnectState getWifiState(final String serviceId) {
-        return ConnectState.NONE;
-    }
-
-    /**
-     * Bluetoothの接続状態を取得する.
-     * 
-     * @param serviceId サービスID
-     * @return Bluetoothの接続状態
-     * @see ConnectState
-     */
-    protected ConnectState getBluetoothState(final String serviceId) {
-        return ConnectState.NONE;
-    }
-
-    /**
-     * NFCの接続状態を取得する.
-     * 
-     * @param serviceId サービスID
-     * @return NFCの接続状態
-     * @see ConnectState
-     */
-    protected ConnectState getNFCState(final String serviceId) {
-        return ConnectState.NONE;
-    }
-
-    /**
-     * BLEの接続状態を取得する.
-     * 
-     * @param serviceId サービスID
-     * @return BLEの接続状態
-     * @see ConnectState
-     */
-    protected ConnectState getBLEState(final String serviceId) {
-        return ConnectState.NONE;
     }
 
     @Override
@@ -163,8 +73,6 @@ public abstract class SystemProfile extends DConnectProfile implements SystemPro
 
         if (attribute == null) {
             result = onGetSystem(request, response, serviceId);
-        } else if (attribute.equals(ATTRIBUTE_DEVICE)) {
-            result = onGetDevice(request, response, serviceId);
         } else {
             MessageUtils.setUnknownAttributeError(response);
         }
@@ -227,41 +135,6 @@ public abstract class SystemProfile extends DConnectProfile implements SystemPro
         return true;
     }
 
-    /**
-     * 周辺機器のシステム情報取得リクエストハンドラー.<br/>
-     * 周辺機器のシステム情報取得を提供し、その結果をレスポンスパラメータに格納する。
-     * レスポンスパラメータの送信準備が出来た場合は返り値にtrueを指定する事。
-     * 送信準備ができていない場合は、返り値にfalseを指定し、スレッドを立ち上げてそのスレッドで最終的にレスポンスパラメータの送信を行う事。
-     * このメソッドでは自動的にシステム情報を返信する。返信処理に変更を加えたい場合はオーバーライドし、処理を上書きすること。
-     * 
-     * @param request リクエストパラメータ
-     * @param response レスポンスパラメータ
-     * @param serviceId サービスID
-     * @return レスポンスパラメータを送信するか否か
-     */
-    protected boolean onGetDevice(final Intent request, final Intent response, final String serviceId) {
-
-        // connect
-        Bundle connect = new Bundle();
-        setWifiState(connect, getWifiState(serviceId));
-        setBluetoothState(connect, getBluetoothState(serviceId));
-        setNFCState(connect, getNFCState(serviceId));
-        setBLEState(connect, getBLEState(serviceId));
-        setConnect(response, connect);
-
-        // version
-        setVersion(response, getCurrentVersionName());
-
-        // supports
-        ArrayList<String> profiles = new ArrayList<String>();
-        for (DConnectProfile profile : mProvider.getProfileList()) {
-            profiles.add(profile.getProfileName());
-        }
-        setSupports(response, profiles.toArray(new String[0]));
-        setResult(response, DConnectMessage.RESULT_OK);
-
-        return true;
-    }
     
     // ------------------------------------
     // PUT
@@ -332,7 +205,7 @@ public abstract class SystemProfile extends DConnectProfile implements SystemPro
         setUnsupportedError(response);
         return true;
     }
-    
+
     /**
      * イベント解除リクエストハンドラー.<br/>
      * 指定されたセッションキーに紐づくイベントを全て解除し、その結果をレスポンスパラメータに格納する。
@@ -382,117 +255,6 @@ public abstract class SystemProfile extends DConnectProfile implements SystemPro
      */
     public static void setSupports(final Intent response, final List<String> supports) {
         setSupports(response, supports.toArray(new String[supports.size()]));
-    }
-
-    /**
-     * レスポンスにデバイスの接続状態を設定する.
-     * 
-     * @param response レスポンスパラメータ
-     * @param connect デバイスの接続状態
-     */
-    public static void setConnect(final Intent response, final Bundle connect) {
-        response.putExtra(PARAM_CONNECT, connect);
-    }
-
-    /**
-     * 指定されたパラメータに接続状態を設定する.
-     * 
-     * @param b バンドルオブジェクト
-     * @param key パラメータキー
-     * @param state 状態
-     */
-    private static void setConnectionState(final Bundle b, final String key, final ConnectState state) {
-        // 非対応のものは省略
-        switch (state) {
-        case ON:
-            b.putBoolean(key, true);
-            break;
-        case OFF:
-            b.putBoolean(key, false);
-            break;
-        default:
-            break;
-        }
-    }
-
-    /**
-     * デバイスプラグインの接続状態にWiFiの接続状態を設定する.
-     * 
-     * @param connect デバイスプラグインの接続状態パラメータ
-     * @param connecting true:ON、false:OFF
-     */
-    public static void setWifiState(final Bundle connect, final boolean connecting) {
-        connect.putBoolean(PARAM_WIFI, connecting);
-    }
-
-    /**
-     * デバイスプラグインの接続状態にWiFiの接続状態を設定する.
-     * 
-     * @param connect デバイスプラグインの接続状態パラメータ
-     * @param state 接続状態
-     */
-    public static void setWifiState(final Bundle connect, final ConnectState state) {
-        setConnectionState(connect, PARAM_WIFI, state);
-    }
-
-    /**
-     * デバイスプラグインの接続状態にBluetoothの接続状態を設定する.
-     * 
-     * @param connect デバイスプラグインの接続状態パラメータ
-     * @param connecting true:ON、false:OFF
-     */
-    public static void setBluetoothState(final Bundle connect, final boolean connecting) {
-        connect.putBoolean(PARAM_BLUETOOTH, connecting);
-    }
-
-    /**
-     * デバイスプラグインの接続状態にBluetoothの接続状態を設定する.
-     * 
-     * @param connect デバイスプラグインの接続状態パラメータ
-     * @param state 接続状態
-     */
-    public static void setBluetoothState(final Bundle connect, final ConnectState state) {
-        setConnectionState(connect, PARAM_BLUETOOTH, state);
-    }
-
-    /**
-     * デバイスプラグインの接続状態にNFCの接続状態を設定する.
-     * 
-     * @param connect デバイスプラグインの接続状態パラメータ
-     * @param connecting true:ON、false:OFF
-     */
-    public static void setNFCState(final Bundle connect, final boolean connecting) {
-        connect.putBoolean(PARAM_NFC, connecting);
-    }
-
-    /**
-     * デバイスプラグインの接続状態にNFCの接続状態を設定する.
-     * 
-     * @param connect デバイスプラグインの接続状態パラメータ
-     * @param state 接続状態
-     */
-    public static void setNFCState(final Bundle connect, final ConnectState state) {
-        setConnectionState(connect, PARAM_NFC, state);
-    }
-
-    /**
-     * デバイスプラグインの接続状態にBLEの接続状態を設定する.
-     * 
-     * @param connect デバイスプラグインの接続状態パラメータ
-     * @param connecting true:ON、false:OFF
-     */
-    public static void setBLEState(final Bundle connect, final boolean connecting) {
-        connect.putBoolean(PARAM_BLE, connecting);
-    }
-
-    /**
-     * デバイスプラグインの接続状態にBLEの接続状態を設定する.
-     * 
-     * @param connect デバイスプラグインの接続状態パラメータ
-     * @param state 接続状態
-     */
-    public static void setBLEState(final Bundle connect, final ConnectState state) {
-        setConnectionState(connect, PARAM_BLE, state);
     }
 
     /**
