+++ dConnectDevicePlugin/dConnectDeviceSonySW/src/org/deviceconnect/android/deviceplugin/sw/profile/SWSystemProfile.java
@@ -7,31 +7,20 @@
 package org.deviceconnect.android.deviceplugin.sw.profile;
 
 import org.deviceconnect.android.deviceplugin.sw.setting.SWSettingMainActivity;
+import org.deviceconnect.android.event.EventManager;
+import org.deviceconnect.android.message.MessageUtils;
+import org.deviceconnect.android.profile.SystemProfile;
+import org.deviceconnect.message.DConnectMessage;
 
 import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
-
-import org.deviceconnect.android.event.EventManager;
-import org.deviceconnect.android.message.MessageUtils;
-import org.deviceconnect.android.profile.DConnectProfileProvider;
-import org.deviceconnect.android.profile.SystemProfile;
-import org.deviceconnect.message.DConnectMessage;
 
 /**
  * SonySWデバイスプラグインの{@link SystemProfile}実装.
  * @author NTT DOCOMO, INC.
  */
 public class SWSystemProfile extends SystemProfile {
-
-    /**
-     * コンストラクタ.
-     * 
-     * @param provider プロファイルプロバイダ
-     */
-    public SWSystemProfile(final DConnectProfileProvider provider) {
-        super(provider);
-    }
 
     @Override
     protected Class<? extends Activity> getSettingPageActivity(final Intent request, final Bundle param) {
