+++ dConnectDevicePlugin/dConnectDeviceSphero/src/org/deviceconnect/android/deviceplugin/sphero/profile/SpheroSystemProfile.java
@@ -12,32 +12,21 @@ import java.util.List;
 import org.deviceconnect.android.deviceplugin.sphero.SpheroManager;
 import org.deviceconnect.android.deviceplugin.sphero.data.DeviceInfo;
 import org.deviceconnect.android.deviceplugin.sphero.setting.SettingActivity;
+import org.deviceconnect.android.event.Event;
+import org.deviceconnect.android.event.EventManager;
+import org.deviceconnect.android.message.MessageUtils;
+import org.deviceconnect.android.profile.SystemProfile;
+import org.deviceconnect.message.DConnectMessage;
 
 import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
-
-import org.deviceconnect.android.event.Event;
-import org.deviceconnect.android.event.EventManager;
-import org.deviceconnect.android.message.MessageUtils;
-import org.deviceconnect.android.profile.DConnectProfileProvider;
-import org.deviceconnect.android.profile.SystemProfile;
-import org.deviceconnect.message.DConnectMessage;
 
 /**
  * System プロファイル.
  * @author NTT DOCOMO, INC.
  */
 public class SpheroSystemProfile extends SystemProfile {
-
-    /**
-     * システムプロファイルを生成する.
-     * 
-     * @param provider プロバイダー
-     */
-    public SpheroSystemProfile(final DConnectProfileProvider provider) {
-        super(provider);
-    }
 
     @Override
     protected Class<? extends Activity> getSettingPageActivity(final Intent request, final Bundle param) {
