+++ dConnectDevicePlugin/dConnectDeviceTest/src/org/deviceconnect/android/deviceplugin/test/profile/TestSystemProfile.java
@@ -8,7 +8,6 @@ package org.deviceconnect.android.deviceplugin.test.profile;
 
 import org.deviceconnect.android.event.EventManager;
 import org.deviceconnect.android.message.MessageUtils;
-import org.deviceconnect.android.profile.DConnectProfileProvider;
 import org.deviceconnect.android.profile.SystemProfile;
 import org.deviceconnect.message.DConnectMessage;
 
@@ -26,63 +25,6 @@ public class TestSystemProfile extends SystemProfile {
      * バージョン.
      */
     public static final String VERSION = "1.0";
-    
-    /**
-     * サービスIDをチェックする.
-     * @param serviceId サービスID
-     * @return <code>serviceId</code>がテスト用サービスIDに等しい場合はtrue、そうでない場合はfalse
-     */
-    private boolean checkserviceId(final String serviceId) {
-        return TestServiceDiscoveryProfile.SERVICE_ID.equals(serviceId);
-    }
-
-    /**
-     * サービスIDが空の場合のエラーを作成する.
-     * 
-     * @param response レスポンスを格納するIntent
-     */
-    private void createEmptyserviceId(final Intent response) {
-        MessageUtils.setEmptyServiceIdError(response);
-    }
-
-    /**
-     * デバイスが発見できなかった場合のエラーを作成する.
-     * 
-     * @param response レスポンスを格納するIntent
-     */
-    private void createNotFoundService(final Intent response) {
-        MessageUtils.setNotFoundServiceError(response);
-    }
-
-    /**
-     * コンストラクタ.
-     * @param plugin プラグイン
-     */
-    public TestSystemProfile(final DConnectProfileProvider plugin) {
-        super(plugin);
-    }
-
-    @Override
-    protected boolean onGetDevice(final Intent request, final Intent response, final String serviceId) {
-        
-        if (serviceId == null) {
-            createEmptyserviceId(response);
-        } else if (!checkserviceId(serviceId)) {
-            createNotFoundService(response);
-        } else {
-            setVersion(response, VERSION);
-            setSupports(response, new String[]{});
-            Bundle connect = new Bundle();
-            setWifiState(connect, false);
-            setBluetoothState(connect, false);
-            setNFCState(connect, false);
-            setBLEState(connect, false);
-            setConnect(response, connect);
-            setResult(response, DConnectMessage.RESULT_OK);
-        }
-        
-        return true;
-    }
 
     @Override
     protected boolean onPutWakeup(final Intent request, final Intent response, final String pluginId) {
