+++ dConnectManager/dConnectManagerTest/src/org/deviceconnect/android/profile/intent/test/NormalSystemProfileTestCase.java
@@ -12,9 +12,6 @@ import org.deviceconnect.message.intent.message.IntentDConnectMessage;
 import org.deviceconnect.profile.SystemProfileConstants;
 
 import android.content.Intent;
-import android.content.pm.PackageInfo;
-import android.content.pm.PackageManager;
-import android.content.pm.PackageManager.NameNotFoundException;
 import android.os.Bundle;
 import android.os.Parcelable;
 
@@ -79,39 +76,6 @@ public class NormalSystemProfileTestCase extends IntentDConnectTestCase
     }
 
     /**
-     * デバイスのシステムプロファイルを取得する.
-     * <pre>
-     * 【Intent通信】
-     * Action: GET
-     * Extra: 
-     *     profile=system
-     *     attribute=device
-     * </pre>
-     * <pre>
-     * 【期待する動作】
-     * ・resultに0が返ってくること。
-     * </pre>
-     */
-    public void testGetSystemDevice() {
-        Intent request = new Intent(IntentDConnectMessage.ACTION_GET);
-        request.putExtra(DConnectMessage.EXTRA_SERVICE_ID, getServiceId());
-        request.putExtra(DConnectMessage.EXTRA_PROFILE, SystemProfileConstants.PROFILE_NAME);
-        request.putExtra(DConnectMessage.EXTRA_ATTRIBUTE, SystemProfileConstants.ATTRIBUTE_DEVICE);
-        Intent response = sendRequest(request);
-
-        assertResultOK(response);
-        assertEquals(getCurrentVersionName(),
-                response.getStringExtra(SystemProfileConstants.PARAM_VERSION));
-        String[] supports = response.getStringArrayExtra(SystemProfileConstants.PARAM_SUPPORTS);
-        assertNotNull(supports);
-        Bundle connect = response.getBundleExtra(SystemProfileConstants.PARAM_CONNECT);
-        assertEquals(false, connect.getBoolean(SystemProfileConstants.PARAM_WIFI));
-        assertEquals(false, connect.getBoolean(SystemProfileConstants.PARAM_BLUETOOTH));
-        assertEquals(false, connect.getBoolean(SystemProfileConstants.PARAM_NFC));
-        assertEquals(false, connect.getBoolean(SystemProfileConstants.PARAM_BLE));
-    }
-
-    /**
      * デバイスプラグインの機能を有効にするテストを行う.
      * <pre>
      * 【Intent通信】
@@ -133,24 +97,10 @@ public class NormalSystemProfileTestCase extends IntentDConnectTestCase
         Intent request = new Intent(IntentDConnectMessage.ACTION_PUT);
         request.putExtra(SystemProfileConstants.PARAM_PLUGIN_ID, mTestPluginID);
         request.putExtra(DConnectMessage.EXTRA_PROFILE, SystemProfileConstants.PROFILE_NAME);
-        request.putExtra(DConnectMessage.EXTRA_INTERFACE, SystemProfileConstants.ATTRIBUTE_DEVICE);
+        request.putExtra(DConnectMessage.EXTRA_INTERFACE, SystemProfileConstants.INTERFACE_DEVICE);
         request.putExtra(DConnectMessage.EXTRA_ATTRIBUTE, SystemProfileConstants.ATTRIBUTE_WAKEUP);
         Intent response = sendRequest(request);
         assertResultOK(response);
     }
 
-    /**
-     * AndroidManifest.xmlのversionNameを取得する.
-     * @return versionName
-     */
-    private String getCurrentVersionName() {
-        PackageManager packageManager = getContext().getPackageManager();
-        try {
-            PackageInfo packageInfo = packageManager.getPackageInfo(
-                    getContext().getPackageName(), PackageManager.GET_ACTIVITIES);
-            return packageInfo.versionName;
-        } catch (NameNotFoundException e) {
-            return "Unknown";
-        }
-    }
 }
