+++ dConnectManager/dConnectManagerTest/src/org/deviceconnect/android/profile/restful/test/FailSystemProfileTestCase.java
@@ -139,220 +139,6 @@ public class FailSystemProfileTestCase extends RESTfulDConnectTestCase {
     }
 
     /**
-     * serviceIdを指定せずにデバイスのシステムプロファイルを取得する.
-     * <pre>
-     * 【HTTP通信】
-     * Method: GET
-     * Path: /system/device
-     * </pre>
-     * <pre>
-     * 【期待する動作】
-     * ・resultに1が返ってくること。
-     * </pre>
-     */
-    public void testGetSystemDeviceNoServiceId() {
-        URIBuilder builder = TestURIBuilder.createURIBuilder();
-        builder.setProfile(SystemProfileConstants.PROFILE_NAME);
-        builder.setAttribute(SystemProfileConstants.ATTRIBUTE_DEVICE);
-        try {
-            HttpUriRequest request = new HttpGet(builder.toString());
-            JSONObject root = sendRequest(request);
-            assertResultError(ErrorCode.EMPTY_SERVICE_ID.getCode(), root);
-        } catch (JSONException e) {
-            fail("Exception in JSONObject." + e.getMessage());
-        }
-    }
-
-    /**
-     * serviceIdに空文字を指定してデバイスのシステムプロファイルを取得する.
-     * <pre>
-     * 【HTTP通信】
-     * Method: GET
-     * Path: /system/device?serviceId=
-     * </pre>
-     * <pre>
-     * 【期待する動作】
-     * ・resultに1が返ってくること。
-     * </pre>
-     */
-    public void testGetSystemDeviceEmptyServiceId() {
-        URIBuilder builder = TestURIBuilder.createURIBuilder();
-        builder.setProfile(SystemProfileConstants.PROFILE_NAME);
-        builder.setAttribute(SystemProfileConstants.ATTRIBUTE_DEVICE);
-        builder.addParameter(DConnectProfileConstants.PARAM_SERVICE_ID, "");
-        try {
-            HttpUriRequest request = new HttpGet(builder.toString());
-            JSONObject root = sendRequest(request);
-            assertResultError(ErrorCode.NOT_FOUND_SERVICE.getCode(), root);
-        } catch (JSONException e) {
-            fail("Exception in JSONObject." + e.getMessage());
-        }
-    }
-
-    /**
-     * 存在しないserviceIdを指定してデバイスのシステムプロファイルを取得する.
-     * <pre>
-     * 【HTTP通信】
-     * Method: GET
-     * Path: /system/device?serviceId=123456789
-     * </pre>
-     * <pre>
-     * 【期待する動作】
-     * ・resultに1が返ってくること。
-     * </pre>
-     */
-    public void testGetSystemDeviceInvalidServiceId() {
-        URIBuilder builder = TestURIBuilder.createURIBuilder();
-        builder.setProfile(SystemProfileConstants.PROFILE_NAME);
-        builder.setAttribute(SystemProfileConstants.ATTRIBUTE_DEVICE);
-        builder.addParameter(DConnectProfileConstants.PARAM_SERVICE_ID, "123456789");
-        try {
-            HttpUriRequest request = new HttpGet(builder.toString());
-            JSONObject root = sendRequest(request);
-            assertResultError(ErrorCode.NOT_FOUND_SERVICE.getCode(), root);
-        } catch (JSONException e) {
-            fail("Exception in JSONObject." + e.getMessage());
-        }
-    }
-
-    /**
-     * 未定義のパラメータを指定してデバイスのシステムプロファイルを取得する.
-     * <pre>
-     * 【HTTP通信】
-     * Method: GET
-     * Path: /system/device?serviceId=xxxx&abc=abc
-     * </pre>
-     * <pre>
-     * 【期待する動作】
-     * ・resultに1が返ってくること。
-     * </pre>
-     */
-    public void testGetSystemDeviceUndefinedParameter() {
-        URIBuilder builder = TestURIBuilder.createURIBuilder();
-        builder.setProfile(SystemProfileConstants.PROFILE_NAME);
-        builder.setAttribute(SystemProfileConstants.ATTRIBUTE_DEVICE);
-        builder.addParameter(DConnectProfileConstants.PARAM_SERVICE_ID, getServiceId());
-        builder.addParameter("abc", "abc");
-        try {
-            HttpUriRequest request = new HttpGet(builder.toString());
-            JSONObject resp = sendRequest(request);
-            assertResultOK(resp);
-            assertEquals("1.0", 
-                    resp.getString(SystemProfileConstants.PARAM_VERSION));
-            JSONObject connect = resp.getJSONObject(SystemProfileConstants.PARAM_CONNECT);
-            assertNotNull(connect);
-        } catch (JSONException e) {
-            fail("Exception in JSONObject." + e.getMessage());
-        }
-    }
-
-    /**
-     * serviceIdを2重に指定してデバイスのシステムプロファイルを取得する.
-     * <pre>
-     * 【HTTP通信】
-     * Method: GET
-     * Path: /system/device?serviceId=123456789&serviceId=xxxx
-     * </pre>
-     * <pre>
-     * 【期待する動作】
-     * ・先に定義された属性が優先されること。
-     * ・resultに1が返ってくること。
-     * </pre>
-     */
-    public void testGetSystemDeviceDuplicatedServiceId() {
-        URIBuilder builder = TestURIBuilder.createURIBuilder();
-        builder.setProfile(SystemProfileConstants.PROFILE_NAME);
-        builder.setAttribute(SystemProfileConstants.ATTRIBUTE_DEVICE);
-        builder.addParameter(DConnectProfileConstants.PARAM_SERVICE_ID, "123456789");
-        builder.addParameter(DConnectProfileConstants.PARAM_SERVICE_ID, getServiceId());
-        try {
-            HttpUriRequest request = new HttpGet(builder.toString());
-            JSONObject root = sendRequest(request);
-            assertResultError(ErrorCode.NOT_FOUND_SERVICE.getCode(), root);
-        } catch (JSONException e) {
-            fail("Exception in JSONObject." + e.getMessage());
-        }
-    }
-
-    /**
-     * POSTメソッドでデバイスのシステムプロファイルを取得する.
-     * <pre>
-     * 【HTTP通信】
-     * Method: POST
-     * Path: /system/device?serviceId=123456789&serviceId=xxxx
-     * </pre>
-     * <pre>
-     * 【期待する動作】
-     * ・resultに1が返ってくること。
-     * </pre>
-     */
-    public void testGetSystemDeviceInvalidMethodPost() {
-        URIBuilder builder = TestURIBuilder.createURIBuilder();
-        builder.setProfile(SystemProfileConstants.PROFILE_NAME);
-        builder.setAttribute(SystemProfileConstants.ATTRIBUTE_DEVICE);
-        builder.addParameter(DConnectProfileConstants.PARAM_SERVICE_ID, getServiceId());
-        try {
-            HttpUriRequest request = new HttpPost(builder.toString());
-            JSONObject root = sendRequest(request);
-            assertResultError(ErrorCode.NOT_SUPPORT_ACTION.getCode(), root);
-        } catch (JSONException e) {
-            fail("Exception in JSONObject." + e.getMessage());
-        }
-    }
-
-    /**
-     * PUTメソッドでデバイスのシステムプロファイルを取得する.
-     * <pre>
-     * 【HTTP通信】
-     * Method: PUT
-     * Path: /system/device?serviceId=123456789&serviceId=xxxx
-     * </pre>
-     * <pre>
-     * 【期待する動作】
-     * ・resultに1が返ってくること。
-     * </pre>
-     */
-    public void testGetSystemDeviceInvalidMethodPut() {
-        URIBuilder builder = TestURIBuilder.createURIBuilder();
-        builder.setProfile(SystemProfileConstants.PROFILE_NAME);
-        builder.setAttribute(SystemProfileConstants.ATTRIBUTE_DEVICE);
-        builder.addParameter(DConnectProfileConstants.PARAM_SERVICE_ID, getServiceId());
-        try {
-            HttpUriRequest request = new HttpPut(builder.toString());
-            JSONObject root = sendRequest(request);
-            assertResultError(ErrorCode.UNKNOWN_ATTRIBUTE.getCode(), root);
-        } catch (JSONException e) {
-            fail("Exception in JSONObject." + e.getMessage());
-        }
-    }
-
-    /**
-     * DELETEメソッドでデバイスのシステムプロファイルを取得する.
-     * <pre>
-     * 【HTTP通信】
-     * Method: DELETE
-     * Path: /system/device?serviceId=123456789&serviceId=xxxx
-     * </pre>
-     * <pre>
-     * 【期待する動作】
-     * ・resultに1が返ってくること。
-     * </pre>
-     */
-    public void testGetSystemDeviceInvalidMethodDelete() {
-        URIBuilder builder = TestURIBuilder.createURIBuilder();
-        builder.setProfile(SystemProfileConstants.PROFILE_NAME);
-        builder.setAttribute(SystemProfileConstants.ATTRIBUTE_DEVICE);
-        builder.addParameter(DConnectProfileConstants.PARAM_SERVICE_ID, getServiceId());
-        try {
-            HttpUriRequest request = new HttpDelete(builder.toString());
-            JSONObject root = sendRequest(request);
-            assertResultError(ErrorCode.UNKNOWN_ATTRIBUTE.getCode(), root);
-        } catch (JSONException e) {
-            fail("Exception in JSONObject." + e.getMessage());
-        }
-    }
-    
-    /**
      * GETメソッドでイベント全消去要求を送信する.
      * <pre>
      * 【HTTP通信】
